requirements.txt
streamlit
tensorflow-cpu
numpy
pillow
gdown